move chat javascript to inside webpacker folder
