+++ app/webpacker/packs/chat.js
@@ -1,1 +1,1 @@
-import('index-77d633bb.js');
+require('../javascript/index-77d633bb.js');
